fix replay and typing session persistence
Restore replay persistence and the History viewer; fix PB flags, text length, consistency graphs, streaks and lesson completion counts; add custom-text languages; harden frontend session flows, queueing, timers, sound, settings and accessibility; remove dead IPC commands.

+++ crates/data/tests/db_validation.rs
@@ -407,7 +407,7 @@ fn stress_10k_tests_replay_loading() {
                 test_id,
                 frame_index: f,
                 timestamp_ms: f * 100,
-                position: f as i64,
+                position: f,
                 expected_char: "a".to_string(),
                 typed_char: Some("a".to_string()),
                 correct: true,
